+++ audio_be/src/commonMain/kotlin/PlaybackEngineDispatcher.kt
@@ -127,7 +127,10 @@ class PlaybackEngineDispatcher(
     val masterLatencyMs: Double get() = master.latencyMs
 
     /** Render one block to [out]: sum every engine into the shared mix, then run the master stage. */
-    fun renderBlock(cursorFrame: Int, out: ShortArray) {
+    // NB `cursorFrame` is Double, not Int: it is an ABSOLUTE frame on the backend timeline, which
+    // grows for the life of the backend and overflows Int after ~12.4 h. Exact below 2^53
+    // (~5,950 years at 48 kHz). See RenderClock.cursorFrame. Per-sample offsets stay Int.
+    fun renderBlock(cursorFrame: Double, out: ShortArray) {
         val startMs = context.performanceTimeMs()
         clock.cursorFrame = cursorFrame
         mix.clear()
